+++ src/components/bread-crumbs/constants.ts
@@ -1,4 +1,4 @@
-import type { LinkItems } from '../../pages/not-found-page/types';
+import type { LinkItems } from '../../sources/types/types';
 import { PagePath } from '../../router/enums';
 import { messages } from './messages';
