+++ src/components/header/bottom-header/bottom-header.module.css
@@ -22,5 +22,6 @@
 }
 
 .searchButton {
-  height: 35px;
+  align-items: center;
+  padding: 8px 14px;
 }
