+++ src/components/header/top-header/right-menu/constants.ts
@@ -1,4 +1,4 @@
-import type { LinkItems } from '../types';
+import type { LinkItems } from '../../../../sources/types/types';
 import { PagePath } from '../../../../router/enums';
 import { messages } from './messages';
 import { IconType } from '../../../svg-builder/enums';
