+++ src/pages/not-found-page/constants.ts
@@ -1,4 +1,3 @@
 const PATH_TO_IMG_NOTFOUND = '../../../assets/images/not-found.png';
-const PATH_TO_IMG_OOPS = '../../../assets/images/oops.png';
 
-export { PATH_TO_IMG_NOTFOUND, PATH_TO_IMG_OOPS };
+export { PATH_TO_IMG_NOTFOUND };
