+++ src/pages/not-found-page/messages.ts
@@ -2,4 +2,5 @@ export const messages = {
   errorMessage: '404 Not Found',
   oopsMessage: 'oops!',
   backButton: 'Back To Home',
+  textImageOops: ' Oops! The page you requested was not found!',
 };
