+++ src/pages/not-found-page/not-found-page.tsx
@@ -1,5 +1,5 @@
 import { Button } from '../../components/button/button';
-import { PATH_TO_IMG_NOTFOUND, PATH_TO_IMG_OOPS } from './constants';
+import { PATH_TO_IMG_NOTFOUND } from './constants';
 import { messages } from './messages';
 import styles from './not-found.module.css';
 import { useNavigate } from 'react-router-dom';
@@ -7,6 +7,7 @@ import { ButtonVariants } from '../../components/button/enums';
 import { PagePath } from '../../router/enums';
 import { Wrapper } from '../../components/wrapper/wrapper';
 import { BreadCrumbs } from '../../components/bread-crumbs/bread-crumbs';
+import { navigationLinks } from '../../components/bread-crumbs/constants';
 
 export const NotFoundPage = () => {
   const navigate = useNavigate();
@@ -20,7 +21,7 @@ export const NotFoundPage = () => {
       <div className={styles.errorMessageContainer}>
         <Wrapper className={styles.errorWrapper}>
           <h2 className={styles.errorTitle}>{messages.errorMessage}</h2>
-          <BreadCrumbs />
+          <BreadCrumbs navigationLinks={navigationLinks} />
         </Wrapper>
       </div>
 
@@ -30,12 +31,7 @@ export const NotFoundPage = () => {
           src={PATH_TO_IMG_NOTFOUND}
           alt={messages.errorMessage}
         />
-        <img
-          src={PATH_TO_IMG_OOPS}
-          alt={messages.oopsMessage}
-          className={styles.oopsImg}
-        />
-
+        <p className={styles.textImage}>{messages.textImageOops}</p>
         <Button
           variant={ButtonVariants.primary}
           className={styles.backButton}
