--- src/pages/not-found-page/types.ts
@@ -1,4 +1,0 @@
-export type LinkItems = {
-  to: string;
-  label: string;
-};
